Move report query to backend for scheduling
