+++ packages/app/src/api/browser/repos/pulse.repo.ts
@@ -44,38 +44,6 @@ const getTimeByProjectCategoryAndRange = (
   return query
 }
 
-const getTimeByProjectAndRange = (startDate: string, endDate: string) => {
-  const query = `
-  with get_minutes as 
-  (
-    select project from activities_pulse 
-    where time between '${startDate}' and '${endDate}' 
-    group by  project, strftime('%s', time) / 120
-  ) 
-  select project as name, count() * 2 as minutes 
-    from get_minutes 
-    group by project 
-    order by minutes desc 
-  `
-  return query
-}
-
-const getTimeByEntityAndRange = (startDate: string, endDate: string) => {
-  const query = `
-  with get_minutes as 
-  (
-    select entity from activities_pulse 
-    where time between '${startDate}' and '${endDate}' 
-    group by entity, strftime('%s', time) / 120
-  ) 
-  select entity, count() * 2 as minutes 
-    from get_minutes 
-    group by entity 
-    order by minutes desc 
-  `
-  return query
-}
-
 const getTimeByCategoryAndRange = (startDate: string, endDate: string) => {
   const query = `
   with get_minutes as 
@@ -97,7 +65,5 @@ export {
   getLatestPulses,
   getDeepWork,
   getTimeByProjectCategoryAndRange,
-  getTimeByProjectAndRange,
-  getTimeByEntityAndRange,
   getTimeByCategoryAndRange,
 }
