+++ packages/app/src/api/localServer/keys.ts
@@ -1,0 +1,3 @@
+export const reportKeys = {
+  weeklyScores: (startDate: string) => ['weeklyScores', startDate] as const,
+}
