+++ packages/server/src/types/activities.api.d.ts
@@ -26,4 +26,31 @@ declare namespace CodeClimbers {
     name: string
     minutes: number
   }
+
+  export interface PerProjectTimeOverviewDB {
+    minutes: number
+    name: string
+  }
+
+  export interface PerProjectTimeAndCategoryOverviewDB {
+    category: string
+    minutes: number
+    name: string
+  }
+
+  export interface EntityTimeOverviewDB {
+    entity: string
+    minutes: number
+  }
+
+  export interface CategoryTimeOverviewDB {
+    category: string
+    minutes: number
+  }
+
+  export interface DeepWorkPeriod {
+    startDate: string
+    endDate: string
+    time: number
+  }
 }
