+++ packages/server/src/v1/activities/activities.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common'
 import { CreateWakatimePulseDto } from '../dtos/createWakatimePulse.dto'
 import { PulseRepo } from '../database/pulse.repo'
 import os from 'node:os'
-import dayjs from 'dayjs'
+import dayjs, { Dayjs } from 'dayjs'
 import { TimePeriodDto } from '../dtos/getCategoryTimeOverview.dto'
 import { PageDto, PageMetaDto, PageOptionsDto } from '../dtos/pagination.dto'
 import {
@@ -197,6 +197,191 @@ export class ActivitiesService {
     const pulses = await this.pulseRepo.getAllPulses()
     const csvString = this.convertRecordsToCSV(pulses)
     return Buffer.from(csvString, 'utf-8')
+  }
+
+  getDeepWorkBetweenDates = async (
+    selectedStartDate: Dayjs,
+    selectedEndDate: Dayjs,
+  ): Promise<CodeClimbers.DeepWorkPeriod[]> => {
+    const startDate = selectedStartDate?.startOf('day').toISOString()
+    const endDate = selectedEndDate?.endOf('day').toISOString()
+
+    const records = await this.pulseRepo.getDeepWork(startDate, endDate)
+
+    const periods: CodeClimbers.DeepWorkPeriod[] = []
+    let currentPeriod: CodeClimbers.DeepWorkPeriod | null = null
+
+    const isSameDay = (date1: string, date2: string) => {
+      return new Date(date1).toDateString() === new Date(date2).toDateString()
+    }
+
+    records.forEach((item) => {
+      if (currentPeriod && isSameDay(currentPeriod.startDate, item.flowStart)) {
+        currentPeriod.endDate = item.flowStart
+        currentPeriod.time += item.flowTime
+      } else {
+        if (currentPeriod) {
+          periods.push(currentPeriod)
+        }
+        currentPeriod = {
+          startDate: item.flowStart,
+          endDate: item.flowStart,
+          time: item.flowTime,
+        }
+      }
+    })
+
+    if (currentPeriod) {
+      periods.push(currentPeriod)
+    }
+
+    return periods
+  }
+
+  getProjectsTimeByRangeAndCategory = async (
+    selectedStartDate: Dayjs,
+    selectedEndDate: Dayjs,
+  ): Promise<CodeClimbers.PerProjectTimeAndCategoryOverview> => {
+    const startDate = selectedStartDate?.startOf('day').toISOString()
+    const endDate = selectedEndDate?.endOf('day').toISOString()
+
+    const records = await this.pulseRepo.getTimeByProjectCategoryAndRange(
+      startDate,
+      endDate,
+    )
+
+    return records.reduce((acc, row) => {
+      if (!acc[row.category]) {
+        acc[row.category] = []
+      }
+
+      acc[row.category].push({ name: row.name, minutes: row.minutes })
+
+      return acc
+    }, {} as CodeClimbers.PerProjectTimeAndCategoryOverview)
+  }
+
+  getProjectsTimeByRange = async (
+    selectedStartDate: Dayjs,
+    selectedEndDate: Dayjs,
+  ): Promise<CodeClimbers.PerProjectTimeOverviewDB[]> => {
+    const startDate = selectedStartDate?.startOf('day').toISOString()
+    const endDate = selectedEndDate?.endOf('day').toISOString()
+
+    const records = await this.pulseRepo.getTimeByProjectAndRange(
+      startDate,
+      endDate,
+    )
+
+    return records
+  }
+
+  getSocialMediaTimeByRange = async (
+    selectedStartDate: Dayjs,
+    selectedEndDate: Dayjs,
+  ): Promise<CodeClimbers.EntityTimeOverviewDB[]> => {
+    const startDate = selectedStartDate?.startOf('day').toISOString()
+    const endDate = selectedEndDate?.endOf('day').toISOString()
+
+    const records = await this.pulseRepo.getTimeByEntityAndRange(
+      startDate,
+      endDate,
+    )
+
+    const sites = [
+      'facebook',
+      'instagram',
+      'twitter',
+      'linkedin',
+      'tiktok',
+      'snapchat',
+      'youtube',
+      'twitch',
+      'pinterest',
+    ]
+
+    const socialMediaTimes = records.filter((row) =>
+      sites.some((site) => row.entity.toLowerCase().includes(site)),
+    )
+
+    return socialMediaTimes
+  }
+
+  getGrowthAndMasteryScore = async (
+    selectedStartDate: Dayjs,
+    selectedEndDate: Dayjs,
+  ): Promise<CodeClimbers.EntityTimeOverviewDB[]> => {
+    const startDate = selectedStartDate?.startOf('day').toISOString()
+    const endDate = selectedEndDate?.endOf('day').toISOString()
+    const records = await this.pulseRepo.getTimeByEntityAndRange(
+      startDate,
+      endDate,
+    )
+
+    const sites = [
+      'mozilla',
+      'stackoverflow',
+      'devdocs',
+      'coursera',
+      'udemy',
+      'codeacademy',
+      'theodinproject',
+      'freecodecamp',
+      'daily.dev',
+      'dev.to',
+      'hackernews',
+      'leetcode',
+      'hackerank',
+    ]
+
+    const growthSites = records.filter((row) =>
+      sites.some((site) => row.entity.toLowerCase().includes(site)),
+    )
+    return growthSites
+  }
+
+  getCategoryTimeByRange = async (
+    selectedStartDate: Dayjs,
+    selectedEndDate: Dayjs,
+  ): Promise<CodeClimbers.CategoryTimeOverviewDB[]> => {
+    const startDate = selectedStartDate?.startOf('day').toISOString()
+    const endDate = selectedEndDate?.endOf('day').toISOString()
+
+    const records = await this.pulseRepo.getTimeByCategoryAndRange(
+      startDate,
+      endDate,
+    )
+
+    // merge category 'communication' with 'communicating'
+    const mergedRecords = records
+      .map((row) => {
+        if (row.category === 'communication') {
+          return { ...row, category: 'communicating' }
+        }
+        return row
+      })
+      .reduce((acc, row) => {
+        const existingCategory = acc.find((c) => c.category === row.category)
+        if (existingCategory) {
+          existingCategory.minutes += row.minutes
+        } else {
+          acc.push(row)
+        }
+        return acc
+      }, [] as CodeClimbers.CategoryTimeOverviewDB[])
+
+    return mergedRecords
+  }
+
+  getTotalTimeByRange = async (
+    selectedStartDate: Dayjs,
+    selectedEndDate: Dayjs,
+  ): Promise<number> => {
+    const timeRecords = await this.getCategoryTimeByRange(
+      selectedStartDate,
+      selectedEndDate,
+    )
+    return timeRecords.reduce((acc, row) => acc + row.minutes, 0)
   }
 
   private userAgent() {
